% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ece244-package.R
\name{any_cpp}
\alias{any_cpp}
\title{Return TRUE if any element in a vector is TRUE (C++)}
\usage{
any_cpp(x)
}
\arguments{
\item{x}{logical vector}
}
\description{
Return TRUE if any element in a vector is TRUE (C++)
}
