% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ece244-package.R
\name{cummax_cpp}
\alias{cummax_cpp}
\title{Return the cumulative maximum of the coordinates of a vector (C++)}
\usage{
cummax_cpp(x)
}
\arguments{
\item{x}{numeric vector}
}
\description{
Return the cumulative maximum of the coordinates of a vector (C++)
}
