% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ece244-package.R
\name{cummin_cpp}
\alias{cummin_cpp}
\title{Return the cumulative minimum of the coordinates of a vector (C++)}
\usage{
cummin_cpp(x)
}
\arguments{
\item{x}{numeric vector}
}
\description{
Return the cumulative minimum of the coordinates of a vector (C++)
}
