% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ece244-package.R
\name{cumprod_cpp}
\alias{cumprod_cpp}
\title{Return the cumulative product of the coordinates of a vector (C++)}
\usage{
cumprod_cpp(x)
}
\arguments{
\item{x}{numeric vector}
}
\description{
Return the cumulative product of the coordinates of a vector (C++)
}
