% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ece244-package.R
\name{cumsum2_cpp}
\alias{cumsum2_cpp}
\title{Return the cumulative sum of the coordinates of a vector (C++)}
\usage{
cumsum2_cpp(x, na_rm = FALSE)
}
\arguments{
\item{x}{numeric vector}

\item{na_rm}{logical. Should missing values (including \code{NaN}) be removed?}
}
\description{
Return the cumulative sum of the coordinates of a vector (C++)
}
