% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ece244-package.R
\name{cumsum_cpp}
\alias{cumsum_cpp}
\title{Return the cumulative sum of the coordinates of a vector (C++)}
\usage{
cumsum_cpp(x)
}
\arguments{
\item{x}{numeric vector}
}
\description{
Return the cumulative sum of the coordinates of a vector (C++)
}
