% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ece244-package.R
\name{cumsum_r}
\alias{cumsum_r}
\title{Return the cumulative sum of the coordinates of a vector (R)}
\usage{
cumsum_r(x)
}
\arguments{
\item{x}{numeric vector}
}
\description{
Return the cumulative sum of the coordinates of a vector (R)
}
