% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ece244-package.R
\name{diff_cpp}
\alias{diff_cpp}
\title{Return the n-th order lagged differences of the coordinates of a vector (C++)}
\usage{
diff_cpp(x, lag = 1)
}
\arguments{
\item{x}{numeric vector}

\item{lag}{integer}
}
\description{
Return the n-th order lagged differences of the coordinates of a vector (C++)
}
