% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ece244-package.R
\name{mean_r}
\alias{mean_r}
\title{Return the mean of the coordinates of a vector (R)}
\usage{
mean_r(x)
}
\arguments{
\item{x}{numeric vector}
}
\description{
Return the mean of the coordinates of a vector (R)
}
