% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ece244-package.R
\name{range_cpp}
\alias{range_cpp}
\title{Return the range of the coordinates of a vector (C++)}
\usage{
range_cpp(x)
}
\arguments{
\item{x}{numeric vector}
}
\description{
Return the range of the coordinates of a vector (C++)
}
