% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ece244-package.R
\name{rmse2_cpp}
\alias{rmse2_cpp}
\title{Return the root mean square error (C++)}
\usage{
rmse2_cpp(x, x0, na_rm = FALSE)
}
\arguments{
\item{x}{numeric vector}

\item{x0}{numeric value}

\item{na_rm}{logical. Should missing values (including \code{NaN}) be removed?}
}
\description{
Return the root mean square error (C++)
}
