% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ece244-package.R
\name{rmse_cpp}
\alias{rmse_cpp}
\title{Return the root mean square error (C++)}
\usage{
rmse_cpp(x, x0)
}
\arguments{
\item{x}{numeric vector}

\item{x0}{numeric value}
}
\description{
Return the root mean square error (C++)
}
