% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{rmse_r}
\alias{rmse_r}
\title{Return the root mean square error (R)}
\usage{
rmse_r(x, x0)
}
\arguments{
\item{x}{numeric vector}

\item{x0}{numeric value}
}
\description{
Return the root mean square error (R)
}
