% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ece244-package.R
\name{sum_cpp}
\alias{sum_cpp}
\title{Return the sum of the coordinates of a vector (C++)}
\usage{
sum_cpp(x)
}
\arguments{
\item{x}{numeric vector}
}
\description{
Return the sum of the coordinates of a vector (C++)
}
