% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ece244-package.R
\name{var_r}
\alias{var_r}
\title{Return the variance of the coordinates of a vector (R)}
\usage{
var_r(x)
}
\arguments{
\item{x}{numeric vector}
}
\description{
Return the variance of the coordinates of a vector (R)
}
