% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ece244-package.R
\name{which_cpp}
\alias{which_cpp}
\title{Return the index of the TRUE elements in a vector (C++)}
\usage{
which_cpp(x)
}
\arguments{
\item{x}{vector of values}
}
\description{
Return the index of the TRUE elements in a vector (C++)
}
