% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ece244-package.R
\name{which_r}
\alias{which_r}
\title{Return the indexes of the TRUE elements in a vector (R)}
\usage{
which_r(x)
}
\arguments{
\item{x}{vector of values}
}
\description{
Return the indexes of the TRUE elements in a vector (R)
}
