% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ece244-package.R
\name{which_cpp}
\alias{which_cpp}
\title{Return the indexes of the elements of a vector that are TRUE (C++)}
\usage{
which_cpp(x, y)
}
\arguments{
\item{x}{predicate function}

\item{y}{vector of values}
}
\description{
Return the indexes of the elements of a vector that are TRUE (C++)
}
