% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{which_r}
\alias{which_r}
\title{Return the index of the first TRUE element in a vector (R)}
\usage{
which_r(x, y)
}
\arguments{
\item{x}{predicate function}

\item{y}{vector of values}
}
\description{
Return the index of the first TRUE element in a vector (R)
}
